refact: move check and update expiration dates and repositories json generation to apis
