+++ taf/tools/metadata/__init__.py
@@ -1,5 +1,7 @@
 import click
 from taf.api.roles import add_signing_key as add_roles_sining_key
+from taf.api.metadata import update_metadata_expiration_date, check_expiration_dates as check_metadata_expiration_dates
+import taf.developer_tool as developer_tool
 from taf.constants import DEFAULT_RSA_SIGNATURE_SCHEME
 from taf.utils import ISO_DATE_PARAM_TYPE as ISO_DATE
 import datetime
@@ -70,7 +72,7 @@ def attach_to_group(group):
             snapshot will expire on 2022-07-28
             root will expire on 2022-08-19
         """
-        developer_tool.check_expiration_dates(repo_path=path, interval=interval, start_date=start_date)
+        check_metadata_expiration_dates(repo_path=path, interval=interval, start_date=start_date)
 
     @metadata.command()
     @click.argument("path")
@@ -105,5 +107,5 @@ def attach_to_group(group):
         if not len(role):
             print("Specify at least one role")
             return
-        developer_tool.update_metadata_expiration_date(path, role, interval, keystore,
-                                                       scheme, start_date, no_commit)
+        update_metadata_expiration_date(path, role, interval, keystore,
+                                        scheme, start_date, no_commit)
