+++ taf/tools/targets/__init__.py
@@ -1,6 +1,11 @@
 import click
 import taf.developer_tool as developer_tool
-from taf.api.targets import list_targets, add_target_repo, remove_target_repo
+from taf.api.targets import (
+    list_targets,
+    add_target_repo,
+    remove_target_repo,
+    generate_repositories_json as geenerate_targets_repositroies_json
+)
 from taf.constants import DEFAULT_RSA_SIGNATURE_SCHEME
 from taf.exceptions import TAFError
 
@@ -39,6 +44,81 @@ def attach_to_group(group):
         """
         custom = {ctx.args[i][2:]: ctx.args[i+1] for i in range(0, len(ctx.args), 2)}
         add_target_repo(auth_path, target_path, target_name, role, library_dir, keystore, custom)
+
+
+
+    @repo.command()
+    @click.argument("path")
+    @click.option("--library-dir", default=None, help="Directory where target repositories and, "
+                  "optionally, authentication repository are located. If omitted it is "
+                  "calculated based on authentication repository's path. "
+                  "Authentication repo is presumed to be at library-dir/namespace/auth-repo-name")
+    @click.option("--namespace", default=None, help="Namespace of the target repositories. "
+                  "If omitted, it will be assumed that namespace matches the name of the "
+                  "directory which contains the authentication repository")
+    @click.option("--targets-rel-dir", default=None, help="Directory relative to which "
+                  "urls of the target repositories are calculated. Only useful when "
+                  "the target repositories do not have remotes set")
+    @click.option("--custom", default=None, help="A dictionary containing custom "
+                  "targets info which will be added to repositories.json")
+    @click.option("--use-mirrors", is_flag=True, help="Whether to generate mirrors.json or not")
+    def generate_repositories_json(path, library_dir, namespace, targets_rel_dir, custom, use_mirrors):
+        """
+        Generate repositories.json. This file needs to be one of the authentication repository's
+        target files or the updater won't be able to validate target repositories.
+        repositories.json is generated by traversing through all targets and adding an entry
+        with the namespace prefixed name of the target repository as its key and the
+        repository's url and custom data as its value.
+
+        Target repositories are expected to be inside a directory whose name is equal to the specified
+        namespace and which is located inside the root directory. If root directory is E:\\examples\\root
+        and namespace is namespace1, target repositories should be in E:\\examples\\root\\namespace1.
+        If the authentication repository and the target repositories are in the same root directory and
+        the authentication repository is also directly inside a namespace directory, then the common root
+        directory is calculated as two repositories up from the authentication repository's directory.
+        Authentication repository's namespace can, but does not have to be equal to the namespace of target,
+        repositories. If the authentication repository's path is E:\\root\\namespace\\auth-repo, root
+        directory will be determined as E:\\root. If this default value is not correct, it can be redefined
+        through the --library-dir option. If the --namespace option's value is not provided, it is assumed
+        that the namespace of target repositories is equal to the authentication repository's namespace,
+        determined based on the repository's path. E.g. Namespace of E:\\root\\namespace2\\auth-repo
+        is namespace2.
+
+        The url of a repository corresponds to its git remote url if set and to its location on the file
+        system otherwise. Test repositories might not have remotes. If targets-rel-dir is specified
+        and a repository does not have remote url, its url is calculated as a relative path to the
+        repository's location from this path. There are two options of defining urls:
+        1. Directly specifying all urls of all repositories directly inside repositories.json
+        2. Using a separate mirrors.json file. This file will be generated only if use-mirrors flag is provided.
+        The mirrors file consists of a list of templates which contain namespace (org name) and repo name as arguments.
+        E.g. "https://github.com/{org_name}/{repo_name}.git". If a target repository's namespaced name
+        is namespace1/target_repo1, its url will be calculated by replacing {org_name} with namespace1
+        and {repo_name} with target_repo1 and would result in "https://github.com/namespace1/target_repo1.git"
+
+        While urls are the only information that the updater needs, it is possible to add
+        any other data using the custom option. Custom data can either be specified in a .json file
+        whose path is provided when calling this command, or directly entered. Keys is this
+        dictionary are names of the repositories whose custom data should be set and values are
+        custom data dictionaries. For example:
+
+        \b
+        {
+            "test/html-repo": {
+                "type": "html"
+            },
+            "test/xml-repo": {
+                "type": "xml"
+            }
+        }
+
+        Note: this command does not automatically register repositories.json as a target file.
+        It is recommended that the content of the file is reviewed before doing so manually.
+        """
+        geenerate_targets_repositroies_json(
+            path, library_dir, namespace, targets_rel_dir, custom, use_mirrors
+        )
+
+
 
     @targets.command()
     @click.argument("repo_path")
